requirements.txt
alembic==1.14.1
bcrypt==4.2.1
black==23.3.0
blinker==1.9.0
cachelib==0.9.0
click==8.1.3
colorama==0.4.6
Deprecated==1.2.18
ecdsa==0.19.0
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.0
Flask-Cors==3.0.10
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.10.1
flask-marshmallow==0.15.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flask-swagger==0.2.14
flask-swagger-ui==4.11.1
graphviz==0.20.3
greenlet==3.1.1
gunicorn==21.2.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
limits==4.0.1
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mypy-extensions==1.0.0
mysql-connector-python==8.0.26
ordered-set==4.1.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
pluggy==1.5.0
protobuf==5.29.3
psycopg2==2.9.9
psycopg2-binary==2.9.10
pyasn1==0.6.1
Pygments==2.19.1
PyJWT==2.10.1
pytest==8.3.4
python-dotenv==1.0.1
python-jose==3.3.0
PyYAML==6.0.2
redis==5.2.1
rich==13.9.4
rsa==4.9
six==1.17.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
waitress==3.0.2
Werkzeug==3.1.3
wrapt==1.17.2
